+++ docs/source/conf.py
@@ -53,7 +53,7 @@ master_doc = 'contents'
 
 # General information about the project.
 project = u'SNMP SMI compiler'
-copyright = u'2015-2017, Ilya Etingof <etingof@gmail.com>'
+copyright = u'2015-2018, Ilya Etingof <etingof@gmail.com>'
 author = u'Ilya Etingof <etingof@gmail.com>'
 
 # The version info for the project you're documenting, acts as replacement for